requirements.txt
requests == 2.32.3
pyJWT == 2.10.1
pandas == 2.2.3
authlib == 1.4.0
openpyxl == 3.1.5
python-dotenv